refactor: move out the types folder && fix imports
