+++ src/db/__init__.py
@@ -1,8 +1,4 @@
-from . import types, methods, client
+from . import methods, client
 
 
-__all__ = [
-    "types",
-    "methods",
-    "client",
-]
+__all__ = ["methods", "client"]
