+++ src/routers/organizations.py
@@ -1,7 +1,8 @@
 from fastapi import APIRouter, Depends
 
-from db.methods import methods
-from db.types import types, RQ, RS
+from db import methods
+from src import types
+from src.types import RQ, RS
 from utils.auth.auth import get_current_user
 from utils.response import ErrorCodes, ErrorResponse, SuccessfulResponse
