+++ src/types/__init__.py
@@ -1,5 +1,5 @@
-from . import types
+from .types import *
 from .requests import RQ
 from .responses import RS
 
-__all__ = ["types", "RQ", "RS"]
+__all__ = ["RQ", "RS"]
